docs: update license file
This commit changes the license file template to the most recent version
of the BSD 3-Clause License. This way GitHub is be able to recognize it.
The changes are mostly in spacing and formatting, e.g. the conditions
are numbered. It replaces "by" in copyright notice with a comma. The
disclaimer uses copyright "holder" vs. copyright "owner". An addition
it also adds afuetterer to the copyright notice.

Fixes: #17